change all view according to index page.

+++ lib/utils/widget_helper.dart
@@ -30,7 +30,7 @@ Map<String, String> statusIconMap = {
   
   // calculate amount for big card table
   String calcuteAmount(String amount, String price) {
-    return (int.parse(amount) * int.parse(price)).toString();
+    return (double.parse(amount) * double.parse(price)).toString();
   }
 
 String truncateToTwoWords(String text) {
@@ -41,32 +41,3 @@ String truncateToTwoWords(String text) {
   }
   return text;
 }
-
-// class ScreenSize {
-//    MediaQueryData ?mediaQueryData;
-//    double ?screenWidth;
-//    double ?screenHeight;
-//    double ?widthLarge;
-//    double ?heightLarge;
-
-
-//   //  ScreenSize( {
-//   //   Key? key,
-//   //   this.mediaQueryData,
-//   //   this.screenWidth,
-//   //   this.screenHeight,
-//   //   this.widthLarge,
-//   //   this.heightLarge,
-//   // });
-
-
-//   void init(BuildContext context) {
-//     mediaQueryData = MediaQuery.of(context);
-//     screenWidth = mediaQueryData!.size.width;
-//     screenHeight = mediaQueryData!.size.height;
-//     widthLarge = screenWidth! / 1280 ;
-//     heightLarge = screenHeight! / 720 ;
-
-//   }
-
-// }
